Minor refactoring of userdata.go
